server/requirements.txt
# LLM Platform Dependencies
# FastAPI and related
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# YAML config
pyyaml>=6.0.1

# HTTP client
httpx>=0.26.0
requests>=2.31.0

# Async
python-multipart>=0.0.6

# JSON handling
orjson>=3.9.0

# CORS
starlette>=0.35.0

# Authentication (optional)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate limiting
slowapi>=0.1.9

# Logging
loguru>=0.7.0

# Environment variables
python-dotenv>=1.0.0

# WebSocket
websockets>=12.0

# Token counting
tiktoken>=0.5.0

# System monitoring
psutil>=5.9.0
